Moved navbar back to its own component
